+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 
 Items starting with `DEPRECATE` are important deprecation notices.
 
+## 3.0.1 (2019-05-21)
+
+### Build process
+
++ Fix empty SERVER_NAME php value when localhost is using
++ Fix issue #27 : related to applicative user uid's change
++ Fix #28 Add missing php packages and libs
+
+## 3.0.0 (2019-03-10)
+
+### Build process
+
+- Use the official PHP image as FROM
+- Separate image variant into dedicated dockerfiles
+
+### Image
+
+- Replace start.sh by entrypoint
+- Remove cron job from supervisord
+- Add /opt/scripts/cronwrapper.py to format cron output
+- Remove support of GLPI_PLUGINS to install plugins, you must use GLPI_INSTALL_PLUGINS
+
 ## 2.4.2 (2018-12-16)
 
 ### Image